chore: add DI token
