+++ projects/kit/components/confirm/index.ts
@@ -1,2 +1,3 @@
 export * from './confirm.component';
+export * from './confirm.options';
 export * from './confirm.service';
